README.md
The Modeling DNA Mutations using Poisson Distribution project simulates and visualizes the occurrence of mutations along a DNA strand based on a Poisson distribution. The Poisson distribution is widely used to model rare events, making it an ideal choice for simulating DNA mutations, which occur infrequently along a long strand of genetic material.